also encode the content of the parts
